[chore] Move all TS tests to the new location
