+++ packages/typescript-tests/src/theming1.tsx
@@ -8,7 +8,7 @@
  */
 
 /* eslint-disable no-unused-vars */
-
+import * as React from 'react';
 import * as stylex from '@stylexjs/stylex';
 import type {
   TokensFromVarGroup,
