+++ packages/typescript-tests/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "$schema": "https://json.schemastore.org/tsconfig",
   "compilerOptions": {
+    "jsx": "react",
     "skipLibCheck": true,
     "moduleDetection": "force",
     "isolatedModules": true,
